Crate ghp

ghp-rs was was created for the easy of moving any folder to any branch.

Enums

Error
Error just neatly wraps the other errors used in the import.

Functions

import_dir
import_dir takes a directory, a branch and a message and will create a commit on that branch with the contents of the directory.

Type Definitions

Result